+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ unreleased
 ----------
 
 * Remove "(unreleased)"
+* Add separator ►
+* Remove what follows [something]
 
 0.2.0 (2020-03-07)
 ------------------